+++ app/api/items.py
@@ -9,14 +9,6 @@ from . import api
 @api.route('/categories/<int:category_id>/items', methods=['GET','POST'])
 @jwt_required()
 def item(category_id):
-     def validate_category(category_id):
-         category = Category.query.filter_by(id=category_id, user_id=user.id).first()
-
-         if not category:
-             return error(data={"error":"Category is not existing."}, status_code=404)
-
-         return category
-
      try:
          user = User.query.filter_by(username=get_jwt_identity()).first()
 
@@ -25,7 +17,10 @@ def item(category_id):
              name = content['name']
              balance = content['balance']
              
-             category = validate_category(category_id)
+             category = Category.query.filter_by(id=category_id, user_id=user.id).first()
+
+             if not category:
+                 return error(data={"error":"Category is not existing."}, status_code=404)
 
              item = Item.query.filter_by(name=name, category_id=category.id).first()
 
@@ -41,7 +36,10 @@ def item(category_id):
              return success(data=data, status_code=201)
 
          if request.method == 'GET':
-             category = validate_category(category_id)
+             category = Category.query.filter_by(id=category_id, user_id=user.id).first()
+
+             if not category:
+                 return error(data={"error":"Category is not existing."}, status_code=404)
 
              item_schema = ItemSchema(many=True)
              items = Item.query.filter_by(category_id=category.id).all()
@@ -56,19 +54,6 @@ def item(category_id):
 @api.route('/categories/<int:category_id>/items/<int:item_id>', methods=['GET','PUT','DELETE'])
 @jwt_required()
 def get_update_delete_item(category_id, item_id):
-     def validate_category_item(category_id, item_id):
-         category = Category.query.filter_by(id=category_id, user_id=user.id).first()
-
-         if not category:
-             return error(data=category_error_message, status_code=404)
-
-         item = Item.query.filter_by(category_id=category.id, id=item_id).first()
-
-         if not item:
-             return error(data=item_error_message, status_code=404)
-
-         return item
-
      try:
          user = User.query.filter_by(username=get_jwt_identity()).first()
          item_schema = ItemSchema()
@@ -76,7 +61,15 @@ def get_update_delete_item(category_id, item_id):
          item_error_message = {"error":"Item is not existing."}
          
          if request.method == 'GET':
-             item = validate_category_item(category_id, item_id)
+             category = Category.query.filter_by(id=category_id, user_id=user.id).first()
+
+             if not category:
+                 return error(data=category_error_message, status_code=404)
+
+             item = Item.query.filter_by(category_id=category.id, id=item_id).first()
+
+             if not item:
+                 return error(data=item_error_message, status_code=404)
 
              data = item_schema.dump(item)
 
@@ -88,7 +81,15 @@ def get_update_delete_item(category_id, item_id):
              activity = content['activity']
              balance = content['balance']
 
-             item = validate_category_item(category_id, item_id)
+             category = Category.query.filter_by(id=category_id, user_id=user.id).first()
+
+             if not category:
+                 return error(data=category_error_message, status_code=404)
+
+             item = Item.query.filter_by(category_id=category.id, id=item_id).first()
+
+             if not item:
+                 return error(data=item_error_message, status_code=404)
 
              item.update(name, activity, balance)
 
@@ -97,7 +98,15 @@ def get_update_delete_item(category_id, item_id):
              return success(data=data)
 
          if request.method == 'DELETE':
-             item = validate_category_item(category_id, item_id)
+             category = Category.query.filter_by(id=category_id, user_id=user.id).first()
+
+             if not category:
+                 return error(data=category_error_message, status_code=404)
+
+             item = Item.query.filter_by(category_id=category.id, id=item_id).first()
+
+             if not item:
+                 return error(data=item_error_message, status_code=404)
              
              item.delete()
